Researched how TESTopic handles INFOs, with a focus on TIFC
Class definition for TESTopic and some methods identified, focusing on methods involved in loading a topic's infos. Goal was to determine whether an incorrect DIAL/TIFC subrecord would cause problems with loading infos; research indicates that Bethesda is careful to reallocate a topic's info list as needed in case TIFC is incorrect.