Outfit research: found armor node teardown code
Found the function that the game uses to tear down existing armor 3D on a character model. If I want to improve Skyrim Outfit System by having it directly tamper with character/armor 3D, then the remaining information that I need is:

What is ActorWeightData::unk548? At this point it appears to be "old" body part data, i.e. data and nodes that need to be torn down.

How do we replicate the various "fixup" tasks performed on a newly-created armor node, such as interpolation between weight models and the application of any texture swaps? We know where this code is; I just need to devise some sort of test-case.