Actor decapitation research
Eventual goal of this line of research is to figure out how to sever arbitrary nodes/limbs from actors. Sufficient knowledge both in this field and in attack processing could allow me to reimplement Fallout-style dismemberment.

So far, I have found a static array in the game's code which is used to hold an actor's head and all descendant 3D nodes during decapitation processing. From what I've seen, large portions of the decapitation graphics handling are specific to the head, referring to BGSBodyPartData (i.e. there's no way we'll ever be able to just get this process to work on arbitrary limbs; we'd need to gain a good understanding of it and replicate the parts that aren't head-specific).