Research into active effects
Discovered a global timer system used for Actor::AdvanceTime.

Identified some addiction-related code in ActiveEffect, near the condition update interval handling.

Added a BGSMessage definition with a more correct definition for the message buttons. The type and layout of the condition struct are unknown, though I've found the member to call to check conditions against a given subject and target.