Crafting menu research
Did some research into crafting menus, in the hopes of creating a DLL that will allow the crafting menus to pull from nearby unowned containers in addition to the player's inventory. (That would be a godsend for crafting within player homes.) I have enough information to make a viable attempt at this for alchemy specifically.

Other findings include functions that may be usable for popping up a message box from C++, and several functions for working with inventories.