Investigation into DialogueMenu
Did some research into DialogueMenu, with the goal of being able to force the menu to display at arbitrary times and tamper with its displayed text, to co-opt it as an improvised UI widget. (The menu handles both player dialogue and ambient subtitles. Cobb Positioner uses it as a UI widget using scripts and a little-understood engine glitch; I'd like to have more control over what I can display.)

As part of the above research, we now have an FxResponseArgs class that should be possible to instantiate whenever we actually know the VTBL location for a given argument count. However, it's untested.

The MenuTopicManager singleton should be the focus of any further research into dialogue subtitle UI.