Research into 3D generation
Finished decode of TESObjectREFR::Unk_6C, and decoded Actor's override. Decoded TESBoundObject::Unk_40. Investigated Inventory3DManager and Singleton01B2E998.

So the thing I've been researching is this: appending NIFs to the current 3D tree is enough to get it to render, but not enough to get it to play NetImmerse animations or to be part of the physics simulation. Examination of the Inventory3DManager and a few other places makes it clear that full 3D functionality is, for whatever reason, inextricable from TESObjectREFR behavior, to the point where Bethesda actually has to create a temporary TESObjectREFR (and flag it to never be written to the save) just to render item and spell models in the UI. And yes, I'm pretty sure that TESObjectREFR still wastes a ref handle, though at least it doesn't waste a form ID.

I'm... kind of astonished, really.